Mark sakaki's repositories as static
Sakaki has stopped maintaining `gentoo-b3` and `sakaki-tools` as
of October 31, 2020. Due to Gentoo expiring EAPI 4 and changes to
eclass includes, permanent errors are now starting to appear on
several ebuilds inside these two repositories. To eliminate these
errors the offending packages need to be removed and attempting to
sync the repository would then result in new errors with eix tools.